K8SPSMDB-683: Fix setting and changing PITR compression variables
The operator was setting the proper values but PBM needs disabling and enabling PITR for new values to take effect.

+++ pkg/apis/psmdb/v1/psmdb_types.go
@@ -688,7 +688,7 @@ type PITRSpec struct {
 	Enabled          bool                `json:"enabled,omitempty"`
 	OplogSpanMin     numstr.NumberString `json:"oplogSpanMin,omitempty"`
 	CompressionType  pbm.CompressionType `json:"compressionType,omitempty"`
-	CompressionLevel *int64              `json:"compressionLevel,omitempty"`
+	CompressionLevel *int                `json:"compressionLevel,omitempty"`
 }
 
 func (p PITRSpec) Disabled() PITRSpec {
